// src/pages/UserSettings/labels.d.ts
import type { ILabels } from 'src/common/Form/types';
export declare const buttons: {
    changeEmail: string;
    changePassword: string;
    deleteLink: {
        message: string;
        text: string;
    };
    impact: {
        create: string;
        edit: string;
        expandOpen: string;
        expandClose: string;
        save: string;
    };
    guidelines: string;
    link: {
        add: string;
        type: string;
    };
    deleteAccount: {
        cancel: string;
        confirm: string;
    };
    map: string;
    notifications: string;
    editPin: string;
    removePin: string;
    save: string;
    success: string;
    submit: string;
    submitNewEmail: string;
    submitNewPassword: string;
};
export declare const fields: ILabels;
export declare const form: {
    defaultError: string;
    saveSuccess: string;
    deleteAccountTitle: string;
    deleteAccountMessage: string;
    deleteAccountLabel: string;
};
export declare const headings: {
    accountSettings: string;
    changeEmail: string;
    changePassword: string;
    createProfile: string;
    editProfile: string;
    focus: string;
    images: string;
    infos: string;
    map: {
        description: string;
        addPinTitle: string;
        yourPinTitle: string;
        existingPinLabel: string;
    };
    visitors: string;
    workspace: {
        description: string;
        title: string;
    };
};
export declare const notificationForm: {
    loading: string;
    successfulSave: string;
};
export declare const mapForm: {
    confirmDeletePin: string;
    descriptionMember: string;
    descriptionSpace: string;
    loading: string;
    locationLabel: string;
    needsChanges: string;
    noLocationLabel: string;
    successfulSave: string;
    successfulDelete: string;
};
export declare const missingData = "Do you have impact data for this year?";
export declare const inCompleteProfile = "In order to add yourself to the map, you need to complete your profile";